client: add core API test cases (#25)
* client: add core api test

* client: fix argument validation according to test

* Assert error messages after unwrap

* Fix flields of get node info response to proper size

* Fix size of jre_available_processors

+++ iota-client/core/get_node_info.rs
@@ -30,16 +30,16 @@ pub struct GetNodeInfoResponse {
     app_version: String,
     /// Number of threads IRI is using
     #[serde(rename = "jreAvailableProcessors")]
-    jre_available_processors: usize,
+    jre_available_processors: u16,
     /// Amount of free memory on IRI node
     #[serde(rename = "jreFreeMemory")]
-    jre_free_memory: usize,
+    jre_free_memory: u64,
     /// Max amount of memory on IRI node
     #[serde(rename = "jreMaxMemory")]
-    jre_max_memory: usize,
+    jre_max_memory: u64,
     /// Total amount of memory on IRI node
     #[serde(rename = "jreTotalMemory")]
-    jre_total_memory: usize,
+    jre_total_memory: u64,
     /// JRE version of IRI node
     #[serde(rename = "jreVersion")]
     jre_version: String,
@@ -48,28 +48,28 @@ pub struct GetNodeInfoResponse {
     latest_milestone: String,
     /// Latest milestone index on IRI node
     #[serde(rename = "latestMilestoneIndex")]
-    latest_milestone_index: usize,
+    latest_milestone_index: u32,
     /// Latest solid subtangle milestone on IRI node
     #[serde(rename = "latestSolidSubtangleMilestone")]
     latest_solid_subtangle_milestone: String,
     /// Latest solid subtangle milestone index on IRI node
     #[serde(rename = "latestSolidSubtangleMilestoneIndex")]
-    latest_solid_subtangle_milestone_index: usize,
+    latest_solid_subtangle_milestone_index: u32,
     /// Milestone start index on IRI node
     #[serde(rename = "milestoneStartIndex")]
-    milestone_start_index: usize,
+    milestone_start_index: u32,
     /// Amount of neighbors connected to IRI node
-    neighbors: usize,
+    neighbors: u16,
     /// Packet queue size on IRI node
     #[serde(rename = "packetsQueueSize")]
-    packets_queue_size: usize,
+    packets_queue_size: u16,
     /// Current time on IRI node (UNIX Seconds),
-    time: usize,
+    time: u64,
     /// Amount of tips on IRI node
-    tips: usize,
+    tips: u32,
     /// Transactions to request on IRI node
     #[serde(rename = "transactionsToRequest")]
-    transactions_to_request: usize,
+    transactions_to_request: u32,
 }
 
 impl GetNodeInfoResponse {
@@ -82,19 +82,19 @@ impl GetNodeInfoResponse {
         &self.app_version
     }
     /// Returns the jre_available_processors attribute
-    pub fn jre_available_processors(&self) -> usize {
+    pub fn jre_available_processors(&self) -> u16 {
         self.jre_available_processors
     }
     /// Returns the jre_free_memory attribute
-    pub fn jre_free_memory(&self) -> usize {
+    pub fn jre_free_memory(&self) -> u64 {
         self.jre_free_memory
     }
     /// Returns the jre_max_memory attribute
-    pub fn jre_max_memory(&self) -> usize {
+    pub fn jre_max_memory(&self) -> u64 {
         self.jre_max_memory
     }
     /// Returns the jre_total_memory attribute
-    pub fn jre_total_memory(&self) -> usize {
+    pub fn jre_total_memory(&self) -> u64 {
         self.jre_total_memory
     }
     /// Returns the jre_version attribute
@@ -106,7 +106,7 @@ impl GetNodeInfoResponse {
         &self.latest_milestone
     }
     /// Returns the latest_milestone_index attribute
-    pub fn latest_milestone_index(&self) -> usize {
+    pub fn latest_milestone_index(&self) -> u32 {
         self.latest_milestone_index
     }
     /// Returns the latest_solid_subtangle_milestone attribute
@@ -114,31 +114,31 @@ impl GetNodeInfoResponse {
         &self.latest_solid_subtangle_milestone
     }
     /// Returns the latest_solid_subtangle_milestone_index attribute
-    pub fn latest_solid_subtangle_milestone_index(&self) -> usize {
+    pub fn latest_solid_subtangle_milestone_index(&self) -> u32 {
         self.latest_solid_subtangle_milestone_index
     }
     /// Returns the milestone_start_index attribute
-    pub fn milestone_start_index(&self) -> usize {
+    pub fn milestone_start_index(&self) -> u32 {
         self.milestone_start_index
     }
     /// Returns the neighbors attribute
-    pub fn neighbors(&self) -> usize {
+    pub fn neighbors(&self) -> u16 {
         self.neighbors
     }
     /// Returns the packets_queue_size attribute
-    pub fn packets_queue_size(&self) -> usize {
+    pub fn packets_queue_size(&self) -> u16 {
         self.packets_queue_size
     }
     /// Returns the time attribute
-    pub fn time(&self) -> usize {
+    pub fn time(&self) -> u64 {
         self.time
     }
     /// Returns the tips attribute
-    pub fn tips(&self) -> usize {
+    pub fn tips(&self) -> u32 {
         self.tips
     }
     /// Returns the transactions_to_request attribute
-    pub fn transactions_to_request(&self) -> usize {
+    pub fn transactions_to_request(&self) -> u32 {
         self.transactions_to_request
     }
 }
